Rename api_factory module to gapic.
