+++ gapic/schema/metadata.py
