+++ gapic/templates/$namespace/$name/__init__.py.j2
