+++ gapic/templates/$namespace/$name_$version/$service/__init__.py.j2
