+++ gapic/templates/$namespace/$name_$version/$service/client.py.j2
