+++ gapic/templates/$namespace/$name_$version/$service/transports/__init__.py.j2
