+++ gapic/templates/$namespace/$name_$version/$service/transports/base.py.j2
