+++ gapic/templates/$namespace/$name_$version/$service/transports/grpc.py.j2
