+++ gapic/templates/$namespace/$name_$version/$service/transports/http.py.j2
