+++ gapic/templates/$namespace/$name_$version/__init__.py.j2
