+++ gapic/templates/$namespace/$name_$version/utils.py.j2
