+++ gapic/templates/setup.py.j2
