+++ gapic/utils/lines.py
