+++ gapic/utils/placeholder.py
