+++ tests/unit/schema/wrappers/test_enums.py
@@ -16,8 +16,8 @@ from typing import Tuple
 
 from google.protobuf import descriptor_pb2
 
-from api_factory.schema import metadata
-from api_factory.schema import wrappers
+from gapic.schema import metadata
+from gapic.schema import wrappers
 
 
 def test_enum_properties():
