+++ tests/unit/schema/wrappers/test_message.py
@@ -19,8 +19,8 @@ import pytest
 
 from google.protobuf import descriptor_pb2
 
-from api_factory.schema import metadata
-from api_factory.schema import wrappers
+from gapic.schema import metadata
+from gapic.schema import wrappers
 
 
 def test_message_properties():
