+++ tests/unit/schema/wrappers/test_method.py
@@ -19,8 +19,8 @@ from google.api import annotations_pb2
 from google.api import signature_pb2
 from google.protobuf import descriptor_pb2
 
-from api_factory.schema import metadata
-from api_factory.schema import wrappers
+from gapic.schema import metadata
+from gapic.schema import wrappers
 
 
 def test_method_types():
