+++ tests/unit/schema/wrappers/test_service.py
@@ -18,8 +18,8 @@ from google.api import annotations_pb2
 from google.api import http_pb2
 from google.protobuf import descriptor_pb2
 
-from api_factory.schema import metadata
-from api_factory.schema import wrappers
+from gapic.schema import metadata
+from gapic.schema import wrappers
 
 
 def test_service_properties():
