+++ tests/unit/utils/test_case.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from api_factory.utils import case
+from gapic.utils import case
 
 
 def test_pascal_to_snake():
