+++ tests/unit/utils/test_filename.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from api_factory.utils import filename
+from gapic.utils import filename
 
 
 def test_to_valid_filename():
